+++ HISTORY.md
@@ -4,6 +4,7 @@
 
 ### 2020-07-23
 * morphic: fixed mouseDown events for touch devices
+* morphic, gui: added separators to list morphs, '~' for the libraries dialog 
 
 ### 2020-07-22
 * morphic, blocks, gui: tweaked block-fading mouse-over